Move serve view tests to their own module and extend to cover invalid ID, signature and filename scenarios
